+++ src/WFU.Bridge/Class1.cs
@@ -1,0 +1,6 @@
+﻿namespace WFU.Bridge;
+
+public class Class1
+{
+
+}